+++ eRecruiter.Api.Client/Requests/Mailer/MailerRequests.cs
@@ -17,4 +17,11 @@ namespace eRecruiter.Api.Client.Requests
         {
         }
     }
+
+    public class SendDoubleOptInMailRequest : PostJsonHttpRequestMessage<MailerResponse>
+    {
+        public SendDoubleOptInMailRequest([NotNull] DoubleOptInParameter parameter) : base("Api/Mailer/DoubleOptIn", parameter)
+        {
+        }
+    }
 }